chore(web): add export lynx-view tests for other projects
